+++ startup/scripts/073-alter largo de nombre de adjuntos.sql
@@ -1,0 +1,6 @@
+ALTER TABLE `draft_metadata_file`
+CHANGE COLUMN `mfi_caption` `mfi_caption` VARCHAR(200) NOT NULL ;
+ALTER TABLE `metadata_file`
+CHANGE COLUMN `mfi_caption` `mfi_caption` VARCHAR(200) NOT NULL ;
+
+UPDATE version SET ver_value = '073' WHERE ver_name = 'DB';